chore: document and split local-first hardening
