+++ Actualist/LocalFirst/Database/BudgetDatabase.swift
@@ -1,45 +1,6 @@
 import Foundation
 import GRDB
 
-private struct EnvelopeCategoryValue {
-    var budgeted: Int = 0
-    var spent: Int = 0
-    var balance: Int = 0
-    var carryover: Bool = false
-}
-
-private struct TransactionBudgetSource {
-    let tableExists: Bool
-    let table: String
-    let account: String
-    let category: String
-    let amount: String
-    let month: String
-    let livePredicate: String
-}
-
-struct ActualSyncDecodedMessage: Equatable, Sendable {
-    let timestamp: String
-    let dataset: String
-    let row: String
-    let column: String
-    let serializedValue: String
-}
-
-struct PendingLocalSyncMessage: Equatable, Sendable {
-    let message: ActualSyncDecodedMessage
-    let baseTimestamp: String
-    let attemptCount: Int
-    let lastError: String?
-}
-
-private enum ActualSyncSQLiteValue {
-    case null
-    case int(Int64)
-    case double(Double)
-    case string(String)
-}
-
 final class BudgetDatabase: @unchecked Sendable {
     let databaseURL: URL
     private let queue: DatabaseQueue
